scripts/requirements.txt
tqdm
numpy
pyyaml
xmltodict
opencv-python
